requirements.txt
actionlib==1.13.2
angles==1.9.13
anyio==3.6.1
apriltag==0.0.16
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.0
bondpy==1.8.6
camera-calibration==1.16.0
camera-calibration-parsers==1.12.0
catkin==0.8.10
catkin-pkg==0.5.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
controller-manager==0.19.5
controller-manager-msgs==0.19.5
cv-bridge==1.16.0
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
diagnostic-analysis==1.11.0
diagnostic-common-diagnostics==1.11.0
diagnostic-updater==1.11.0
distro==1.7.0
docutils==0.18.1
dynamic-reconfigure==1.7.3
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
fonttools==4.33.3
gazebo-plugins==2.9.2
gazebo-ros==2.9.2
gencpp==0.6.5
geneus==3.0.0
genlisp==0.4.18
genmsg==0.5.16
gennodejs==2.0.2
genpy==0.6.15
ghalton==0.6.2
idna==3.3
image-geometry==1.16.0
imageio==2.6.0
importlib-metadata==4.11.3
importlib-resources==5.7.1
interactive-markers==1.12.0
interruptingcow==0.8
ipykernel==6.13.0
ipython==8.3.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.2
joblib==1.1.0
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
json5==0.9.8
jsonschema==4.5.1
jupyter-client==7.3.1
jupyter-core==4.10.0
jupyter-server==1.17.0
jupyterlab==3.4.2
jupyterlab-pygments==0.2.2
jupyterlab-server==2.13.0
kiwisolver==1.4.2
laser-geometry==1.6.7
-e git+git@github.com:HIRO-group/logger.git@fdc2750dbf3278f7b5d76f702644cd6e0068723c#egg=logger
lxml==4.8.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
message-filters==1.15.14
mistune==0.8.4
moveit-core==1.1.9
nbclassic==0.3.7
nbclient==0.6.3
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8
notebook==6.4.11
notebook-shim==0.1.0
-e git+git@github.com:HIRO-group/npm-base.git@151c5cb608859afd66c21a77343d8736d946d74a#egg=npm_base
-e git+git@github.com:HIRO-group/npm-models.git@fa42a8f0389c5b73a52d5b52d54f40df393f2de4#egg=npm_models
-e git+git@github.com:HIRO-group/npm-planning.git@5c64694458cfbe03d736d34e37e80f7e06a22591#egg=npm_planning
numpy==1.22.3
opencv-python==4.2.0.32
packaging==21.3
-e git+git@github.com:HIRO-group/panda-dynamics-model.git@d2afe42f6f6e8ea3ce6b6ea92abdff04d572445d#egg=panda_dynamics_model
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pybind11==2.10.0
pybullet==3.2.5
pybullet-planning==0.5.0
pycparser==2.21
Pygments==2.12.0
pyparsing==3.0.9
pyrealsense2==2.50.0.3812
pyrsistent==0.18.1
python-dateutil==2.8.2
python-qt-binding==0.4.4
pytz==2022.1
PyYAML==5.4.1
pyzmq==22.3.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
recordclass==0.17.2
requests==2.27.1
resource-retriever==1.12.7
-e git+git@github.com:HIRO-group/robot-interface.git@a9893b001714fbf1ab1ea29e57a1034bf642f1f6#egg=robot_interface
-e git+git@github.com:HIRO-group/robot-sim-envs.git@81bd59e6c1b024f0716f6e50921aac61e1bab27d#egg=robot_sim_envs
rosbag==1.15.14
rosboost-cfg==1.15.8
rosclean==1.15.8
roscreate==1.15.8
rosgraph==1.15.14
roslaunch==1.15.14
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.14
rosmake==1.15.8
rosmaster==1.15.14
rosmsg==1.15.14
rosnode==1.15.14
rosparam==1.15.14
rospkg==1.4.0
rospy==1.15.14
rospy-message-converter==0.5.2
rosservice==1.15.14
rostest==1.15.14
rostopic==1.15.14
rosunit==1.15.8
roswtf==1.15.14
rqt-action==0.4.9
rqt-bag==0.5.1
rqt-bag-plugins==0.5.1
rqt-console==0.4.11
rqt-dep==0.4.12
rqt-graph==0.4.14
rqt-gui==0.5.3
rqt-gui-py==0.5.3
rqt-image-view==0.4.16
rqt-launch==0.4.9
rqt-logger-level==0.4.11
rqt-moveit==0.5.10
rqt-msg==0.4.10
rqt-nav-view==0.5.7
rqt-plot==0.4.13
rqt-pose-view==0.5.11
rqt-publisher==0.4.10
rqt-py-common==0.5.3
rqt-py-console==0.4.10
rqt-reconfigure==0.5.5
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.14
rqt-robot-steering==0.5.12
rqt-runtime-monitor==0.5.9
rqt-rviz==0.7.0
rqt-service-caller==0.4.10
rqt-shell==0.4.11
rqt-srv==0.4.9
rqt-tf-tree==0.6.3
rqt-top==0.4.10
rqt-topic==0.4.13
rqt-web==0.4.10
ruckig==0.8.4
rviz==1.14.19
scikit-learn==1.1.0
scipy==1.8.0
Send2Trash==1.8.0
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smach-ros==2.5.0
smclib==1.8.6
sniffio==1.2.0
soupsieve==2.3.2.post1
srdfdom==0.6.3
stack-data==0.2.0
termcolor==1.1.0
terminado==0.13.3
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.5
tf2-kdl==0.7.5
tf2-py==0.7.5
tf2-ros==0.7.5
threadpoolctl==3.1.0
tinycss2==1.1.1
topic-tools==1.15.14
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
trac-ik-python==1.6.6
traitlets==5.2.0
transformations==2021.6.6
typing-extensions==4.2.0
urdfdom-py==0.4.6
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.2
xacro==1.14.13
zipp==3.8.0
